Fix flow sampling threshold value from overflowing.
1.  When a flow-sample is getting dropped, reset its diff statistics. The lost
    statistics will be accounted by normalizing of stats of flows that get
    exported. This reset will reduce the number flows with high statistics and
    prevent threshold from overflowing.
2.  Change the flow threshold value from u32 to u64 field
3.  Also, don’t let u64 threshold value to overflow. If any increment of
    threshold causes its value to overflow, don’t increment/change the value of
    threshold

Closes-Bug: #1655276

+++ src/vnsw/agent/vrouter/flow_stats/flow_stats_collector.h
@@ -110,7 +110,7 @@ public:
     void set_flow_tcp_syn_age_time(uint64_t interval) {
         flow_tcp_syn_age_time_ = interval;
     }
-    uint32_t threshold()  const;
+    uint64_t threshold()  const;
     boost::uuids::uuid rand_gen();
     bool Run();
     bool RunAgeingTask();
